fix(playback): UpsertUserItemData 中 $3 显式转 bigint 避免 int4 溢出
COALESCE($3, 0) 里裸字面量 0 默认 int4,导致整条预处理语句把 playback_position_ticks 参数推断成 int4。tick 值超过 21 亿时 (/Sessions/Playing/Stopped) pgx 编码报 greater than maximum value for int4。INSERT 与 ON CONFLICT 两处 $3 均标注 ::bigint。

+++ internal/models/item.go
@@ -850,9 +850,9 @@ func GetUserItemData(ctx context.Context, pool *pgxpool.Pool, userID, itemID str
 func UpsertUserItemData(ctx context.Context, pool *pgxpool.Pool, userID, itemID string, position *int64, playCount *int32, isFavorite *bool, played *bool) error {
 	_, err := pool.Exec(ctx,
 		`INSERT INTO user_item_data (user_id, item_id, playback_position_ticks, play_count, is_favorite, played, last_played_date)
-		 VALUES ($1::uuid, $2::uuid, COALESCE($3, 0), COALESCE($4, 0), COALESCE($5, false), COALESCE($6, false), NOW())
+		 VALUES ($1::uuid, $2::uuid, COALESCE($3::bigint, 0), COALESCE($4, 0), COALESCE($5, false), COALESCE($6, false), NOW())
 		 ON CONFLICT (user_id, item_id) DO UPDATE SET
-		   playback_position_ticks = COALESCE($3, user_item_data.playback_position_ticks),
+		   playback_position_ticks = COALESCE($3::bigint, user_item_data.playback_position_ticks),
 		   play_count = COALESCE($4, user_item_data.play_count),
 		   is_favorite = COALESCE($5, user_item_data.is_favorite),
 		   played = COALESCE($6, user_item_data.played),
